+++ src/plugins/localstorage/index.ts
@@ -1,3 +1,2 @@
 export { LocalStoragePluginModule } from './localstorage.module';
 export { LocalStoragePlugin } from './localstorage.plugin';
-export { StorageStrategy } from './symbols';